# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Glipz HTTP API
  version: "1.0.0"
  description: |
    Glipz versioned REST API under this document's server base URL.
    
    ## Audience
    BOT operators, integrations, internal scripts, and alternative clients. Anything not described here should be treated as undocumented and may change without notice.
    
    ## Base URL
    Replace the example host with your real API origin. Paths in this document are relative to `/api/v1`. CORS follows the server allowlist.
    
    ## Federation
    Instance-to-instance federation is documented separately (Federation protocol guidelines in the web app at `/federation/guidelines`). This API is mainly for same-instance user actions.
    
    ## Authentication
    - **Login JWT** from `POST /auth/login` (plus MFA when enabled)—same format the web app stores.
    - **Personal access tokens (PATs)** start with `glpat_` and are meant for server-side bots as raw Bearer tokens.
    - **OAuth2**: `client_credentials` and `authorization_code`. Token endpoint: `POST /oauth/token` with `application/x-www-form-urlencoded`.
    
    Always send `Authorization: Bearer <token>` for protected routes. Never put tokens in query strings.
    
...
    
    ## Conventions
    JSON bodies use `Content-Type: application/json` with UTF-8. Media upload uses `multipart/form-data`.
    
...
    
    ## SSE
    Streams such as `/posts/feed/stream` may stay connected; use exponential backoff on reconnect.
    
    ## Disclaimer
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Login JWT, PAT (glpat_...), or OAuth2 access token.
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://api.example.com/api/v1/oauth/token
          scopes: {}
        authorizationCode:
          authorizationUrl: https://api.example.com/developer/oauth/authorize
          tokenUrl: https://api.example.com/api/v1/oauth/token
          scopes: {}
...
paths:
...
  /custom-emojis:
    get:
      tags:
        - Custom emojis
      summary: List enabled custom emojis
      security: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
...
  /public/federation/custom-emoji:
    get:
      tags:
        - Public federation
      summary: Resolve public federated custom emoji
      security: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
...
  /media/remote:
    get:
      tags:
        - Media
      summary: Proxy remote media
      security: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
    head:
      tags:
        - Media
      summary: Remote media metadata
      security: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
...
  /identity/transfers/{sessionID}/media:
    get:
      tags:
        - Identity portability
      summary: Identity transfer media export
      security: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
...
  /me/custom-emojis:
    get:
      tags:
        - Custom emojis
      summary: List my custom emojis
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
    post:
      tags:
        - Custom emojis
      summary: Create my custom emoji
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
  /me/custom-emojis/{emojiID}:
    patch:
      tags:
        - Custom emojis
      summary: Patch my custom emoji
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
    delete:
      tags:
        - Custom emojis
      summary: Delete my custom emoji
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
...
  /posts/feed/custom:
    post:
      tags:
        - Posts
      summary: Authenticated custom feed
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success (or streaming body for SSE)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Too many requests
...